requirements.txt
# Runtime dependencies (install on Databricks cluster via %pip install -r requirements.txt)
yfinance>=0.2.36
newsapi-python>=0.2.7
pyyaml>=6.0

# Development / local testing only
# pyspark and delta-spark are provided by the Databricks runtime — only needed locally
pyspark>=3.4.0
delta-spark>=3.0.0
pandas>=2.0.0
pytest>=7.4.0
pytest-mock>=3.12.0
